52 workout history (#28)
* adding history files, pre destorying the history cache request

* wee refactor

* wip

* history page showing up

* workout details backbone

* wip history

* done, pre testing

* tests

* fix some tests

* add some more tests

+++ SPLYT/SPLYT/Swift Packages/DesignSystem/Sources/DesignSystem/Atomic/Molecules/RoutineTile.swift
@@ -4,13 +4,13 @@ public struct RoutineTile: View {
     @State private var showActionSheet: Bool = false
     private let viewState: RoutineTileViewState
     private let tapAction: () -> Void
-    private let editAction: () -> Void
-    private let deleteAction: () -> Void
+    private let editAction: (() -> Void)?
+    private let deleteAction: (() -> Void)?
     
     public init(viewState: RoutineTileViewState,
                 tapAction: @escaping () -> Void,
-                editAction: @escaping () -> Void,
-                deleteAction: @escaping () -> Void) {
+                editAction: (() -> Void)? = nil,
+                deleteAction: (() -> Void)? = nil) {
         self.viewState = viewState
         self.tapAction = tapAction
         self.editAction = editAction
@@ -47,8 +47,12 @@ public struct RoutineTile: View {
         }
         .confirmationDialog("", isPresented: $showActionSheet, titleVisibility: .hidden) {
             // These just have to be normal buttons
-            Button(Strings.edit) { editAction() }
-            Button(Strings.delete, role: .destructive) { deleteAction() }
+            if let editAction = editAction {
+                Button(Strings.edit) { editAction() }
+            }
+            if let deleteAction = deleteAction {
+                Button(Strings.delete, role: .destructive) { deleteAction() }
+            }
         }
     }
 }
@@ -57,20 +61,17 @@ public struct RoutineTile: View {
 
 public struct RoutineTileViewState: Hashable {
     public let id: String
-    public let historyFilename: String? // Where this routine's history is cached, if it is
     let title: String
     let subtitle: String
     let lastCompletedTitle: String?
     let includeIcon: Bool // For edit and delete actions
     
     public init(id: String,
-                historyFilename: String? = nil,
                 title: String,
                 subtitle: String,
                 lastCompletedTitle: String? = nil,
                 includeIcon: Bool = true) {
         self.id = id
-        self.historyFilename = historyFilename
         self.title = title
         self.subtitle = subtitle
         self.lastCompletedTitle = lastCompletedTitle
